% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iid_models.R
\name{iidJointSmooth}
\alias{iidJointSmooth}
\title{Title}
\usage{
iidJointSmooth(
  Yhat,
  Vhat,
  domain,
  df,
  X = NULL,
  pc_u = c(1, 0.01),
  pc_tau = c(1, 0.01),
  var_tol = 1e-05,
  initf = NULL,
  seed = NULL,
  detailed_output = F,
  chains = 4,
  warmup = 2000,
  iter = 4000,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
