% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_models.R
\name{nb2mat_to_nodes}
\alias{nb2mat_to_nodes}
\title{Create list of nodes from adjacency matrix for spatial modeling}
\usage{
nb2mat_to_nodes(adj_mat)
}
\arguments{
\item{adj_mat}{adjacency matrix}
}
\value{
list of two vectors of nodes that represent edges in graph
}
\description{
Create list of nodes from adjacency matrix for spatial modeling
}
