% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_models.R
\name{prepare_bym2}
\alias{prepare_bym2}
\title{Prepare objects necessary for BYM2 model}
\usage{
prepare_bym2(adj_mat)
}
\arguments{
\item{adj_mat}{adjacency matrix}
}
\value{
list of two vectors of nodes that represent edges in graph and scaling factor necessary for precision matrix
}
\description{
Prepare objects necessary for BYM2 model
}
