% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_joint_logit.R
\name{spatial_joint_logit}
\alias{spatial_joint_logit}
\title{Title}
\usage{
spatial_joint_logit(
  area_vec,
  data_list,
  initf,
  method_name,
  detailed_output = F,
  ...
)
}
\arguments{
\item{initf}{hello}

\item{detailed_output}{}
}
\value{

}
\description{
Title
}
