% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_joint_logit.R
\name{spatialJointLogitVS}
\alias{spatialJointLogitVS}
\title{Title}
\usage{
spatialJointLogitVS(
  Yhat,
  Vhat,
  domain,
  df,
  adj_mat,
  X = NULL,
  pc_u = c(1, 0.01),
  pc_tau = c(1, 0.01),
  var_tol = 1e-05,
  initf = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
