% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_models.R
\name{spatialJointSmoothUnmatched}
\alias{spatialJointSmoothUnmatched}
\title{Unmatched joint smoothing model with BYM2 area effects}
\usage{
spatialJointSmoothUnmatched(
  Yhat,
  Vhat,
  domain,
  na,
  df,
  adj_mat,
  X = NULL,
  pc_u = c(1, 0.01),
  pc_tau = c(1, 0.01),
  var_tol = 1e-05,
  initf = NULL,
  seed = NULL,
  detailed_output = F,
  chains = 4,
  warmup = 2000,
  iter = 4000,
  ...
)
}
\arguments{
\item{Yhat}{vector of direct weighted estimates}

\item{Vhat}{vector of estimated variances of direct weighted estimators}

\item{domain}{vector of domain labels}

\item{na}{vector of sample sizes for areas}

\item{df}{vector of degrees of freedom for areas}

\item{adj_mat}{adjacency matrix}

\item{X}{matrix of area level covariates}

\item{pc_u}{hyperparameters for prior c(t, alpha) such that P(sigma_u > t) = alpha}

\item{pc_tau}{hyperparameters for prior c(t, alpha) such that P(sigma_tau > t) = alpha}

\item{var_tol}{tolerance parameter; all direct estimates with Vhat below var_tol are ignored when smoothing}

\item{initf}{optional stan parameter with starting parameter values}

\item{seed}{random seed}

\item{detailed_output}{if T, return stan object as well as estimates}

\item{chains}{number of chains for Stan}

\item{warmup}{number of warm up samples per chain for Stan}

\item{iter}{total number of samples per chain for Stan}

\item{...}{additional parameters}
}
\value{
A data frame of summaries of the draws from the posterior, including mean, median, variance, and prediction interval
}
\description{
Unmatched joint smoothing model with BYM2 area effects
}
