% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_models.R
\name{summarize_estimates}
\alias{summarize_estimates}
\title{Summarize draws from posterior}
\usage{
summarize_estimates(domain, out_stan, method, logit = F)
}
\arguments{
\item{domain}{vector of domain labels}

\item{out_stan}{stan object from fit}

\item{method}{name of model}

\item{logit}{if T, apply logit transform to draws}
}
\value{
A data frame of summaries of the draws from the posterior, including mean, median, variance, and prediction interval
}
\description{
Summarize draws from posterior
}
